Fix tiny DisplayFactor/TickSize underflowing

+++ apisrc/compat/EfhMsgs.h
@@ -329,10 +329,8 @@ typedef struct {
     #define EfhFutureDefinitionMsg_FIELD_ITER( _x )                         \
                 _x( EfhMsgHeader,    header )                               \
                 _x( EfhSecurityDef,  commonDef )                            \
-                /** Divide by EFH_PRICE_SCALE. */                           \
-                _x( int64_t,         tickSize )                             \
-                /** Divide by EFH_PRICE_SCALE. */                           \
-                _x( int64_t,         displayFactor )
+                _x( float,           tickSize )                             \
+                _x( float,           displayFactor )
         EfhFutureDefinitionMsg_FIELD_ITER( EKA__FIELD_DEF )
 } EfhFutureDefinitionMsg;
